feat: implement DetailsClient component for TCC details display
